refactor : Reservation 생성 도메인으로 옮김
