--- src/main/java/roomescape/common/EndPointPath.java
@@ -1,5 +1,0 @@
-package roomescape.common;
-
-public class EndPointPath {
-    public final static String RESERVATION_API_ENDPOINT_ROOT = "/reservations";
-}
