--- src/main/java/roomescape/common/PagePath.java
@@ -1,6 +1,0 @@
-package roomescape.common;
-
-public class PagePath {
-    public final static String HOME_PAGE_PATH = "/";
-    public final static String RESERVATION_PAGE_PATH = "/reservation";
-}
