+++ src/main/java/roomescape/model/Reservation.java
@@ -1,6 +1,12 @@
 package roomescape.model;
 
-public record Reservation (int id, String name, String date, String time) {
+import roomescape.dto.ReservationDto;
+
+public record Reservation(int id, String name, String date, String time) {
+    public Reservation (int id, ReservationDto reservationDto) {
+        this(id, reservationDto.name(), reservationDto.date(), reservationDto.time());
+    }
+
     public Reservation copy() {
         return new Reservation(this.id, this.name, this.date, this.time);
     }
